[#9] Refactor kb router route & fix test
