+++ main/tests/functional/v1/test_knowledge_base_delete_route.py
@@ -13,7 +13,7 @@ class TestDeleteKnowledgeBase:
         return {"Authorization": f"API-Key {api_key_value}"}
 
     async def test_delete_kb_requires_api_key(
-        self, app, session, client, patch_kb_route_services
+        self, app, session, client, patch_external_services
     ):
         """No API key should return 401"""
         kb = KnowledgeBase(name="Test KB", description="A test KB")
@@ -21,7 +21,7 @@ class TestDeleteKnowledgeBase:
         session.commit()
         kb_id = kb.id
 
-        mock_minio, mock_vector, _, _ = patch_kb_route_services
+        mock_minio = patch_external_services["mock_minio"]
         mock_minio.list_objects.return_value = []
 
         res = await client.delete(
@@ -36,15 +36,17 @@ class TestDeleteKnowledgeBase:
         app,
         session,
         api_key_value,
-        patch_kb_route_services,
+        patch_external_services,
     ):
         kb = KnowledgeBase(name="Test KB", description="A test KB")
         session.add(kb)
         session.commit()
         kb_id = kb.id
 
-        mock_minio, mock_vector, _, _ = patch_kb_route_services
+        mock_minio = patch_external_services["mock_minio"]
         mock_minio.list_objects.return_value = []
+
+        mock_vector = patch_external_services["mock_vector_store"]
 
         response = await client.delete(
             app.url_path_for("v1_delete_knowledge_base", kb_id=kb_id),
@@ -76,14 +78,14 @@ class TestDeleteKnowledgeBase:
         app,
         session,
         api_key_value,
-        patch_kb_route_services,
+        patch_external_services,
     ):
         kb = KnowledgeBase(name="Test KB 2", description="A test KB 2")
         session.add(kb)
         session.commit()
         kb_id = kb.id
 
-        mock_minio, _, _, _ = patch_kb_route_services
+        mock_minio = patch_external_services["mock_minio"]
         mock_minio.list_objects.side_effect = MinioException(
             code="500",
             message="MinIO down",
@@ -100,9 +102,7 @@ class TestDeleteKnowledgeBase:
 
         assert response.status_code == status.HTTP_200_OK
         assert "warnings" in response.json()
-        assert (
-            "Failed to clean up MinIO files" in response.json()["warnings"][0]
-        )
+        assert "MinIO cleanup failed" in response.json()["warnings"][0]
         assert session.query(KnowledgeBase).filter_by(id=kb_id).first() is None
 
     async def test_delete_chroma_error(
@@ -111,15 +111,17 @@ class TestDeleteKnowledgeBase:
         app,
         session,
         api_key_value,
-        patch_kb_route_services,
+        patch_external_services,
     ):
         kb = KnowledgeBase(name="Test KB 3", description="A test KB 3")
         session.add(kb)
         session.commit()
         kb_id = kb.id
 
-        mock_minio, mock_vector, _, _ = patch_kb_route_services
+        mock_minio = patch_external_services["mock_minio"]
         mock_minio.list_objects.return_value = []
+
+        mock_vector = patch_external_services["mock_vector_store"]
         mock_vector.delete_collection.side_effect = Exception("Chroma down")
 
         response = await client.delete(
@@ -130,7 +132,8 @@ class TestDeleteKnowledgeBase:
         assert response.status_code == status.HTTP_200_OK
         assert "warnings" in response.json()
         assert (
-            "Failed to clean up vector store" in response.json()["warnings"][0]
+            "Vector store cleanup failed: Chroma down"
+            in response.json()["warnings"][0]
         )
         assert session.query(KnowledgeBase).filter_by(id=kb_id).first() is None
 
@@ -140,14 +143,14 @@ class TestDeleteKnowledgeBase:
         app,
         session,
         api_key_value,
-        patch_kb_route_services,
+        patch_external_services,
     ):
         kb = KnowledgeBase(name="Test KB 4", description="A test KB 4")
         session.add(kb)
         session.commit()
         kb_id = kb.id
 
-        mock_minio, mock_vector, _, _ = patch_kb_route_services
+        mock_minio = patch_external_services["mock_minio"]
         mock_minio.list_objects.return_value = []
         # simulate unexpected failure
         mock_minio.list_objects.side_effect = RuntimeError("Unexpected error")
